Port subset of dired.c to rust (code review rename dired_rust.rs)
